+++ packages/core/src/index.umd.js
@@ -2,5 +2,4 @@ export { default } from '.'
 export * from '.'
 
 import * as effects from './effects'
-import * as utils from './utils'
-export { effects, utils }
+export { effects }
