--- packages/core/src/utils.js
@@ -1,2 +1,0 @@
-export { SAGA_ACTION, TASK } from '@redux-saga/symbols'
-export { createMockTask, cloneableGenerator } from './internal/utils'
