+++ packages/delay-p/index.d.ts
@@ -1,0 +1,2 @@
+export default function delayP(ms: number): Promise<true>;
+export default function delayP<T>(ms: number, val: T): Promise<T>;
