+++ packages/delay-p/src/index.js
@@ -1,9 +1,9 @@
 import { CANCEL } from '@redux-saga/symbols'
 
-export default function delayP(ms, returnValue = true) {
+export default function delayP(ms, val = true) {
   let timeoutId
   const promise = new Promise(resolve => {
-    timeoutId = setTimeout(resolve, ms, returnValue)
+    timeoutId = setTimeout(resolve, ms, val)
   })
 
   promise[CANCEL] = () => {
