+++ packages/symbols/index.d.ts
@@ -1,0 +1,11 @@
+export const CANCEL: string;
+export const CHANNEL_END_TYPE: string;
+export const IO: string;
+export const MATCH: string;
+export const MULTICAST: string;
+export const SAGA_ACTION: string;
+export const SAGA_LOCATION: string;
+export const SELF_CANCELLATION: string;
+export const TASK: string;
+export const TASK_CANCEL: string;
+export const TERMINATE: string;
